✨ feat(agent): unify agent page headers with a shared breadcrumb
Extract the topic header's inline breadcrumb into a reusable AgentBreadcrumb
(`<AgentName> › <Section>`, agent name links home) and reuse it across the
agent profile (助理档案) and usage & cost (用量与成本) pages so sibling pages
read consistently. Identity/location (breadcrumb) leads the profile header,
separated from the state tags that follow.
