+++ src/routes/(main)/agent/profile/features/Header/index.tsx
@@ -18,6 +18,7 @@ import { useAgentTransferMenuItem } from '@/business/client/hooks/useAgentTransf
 import { useBusinessAgentImportMenuItem } from '@/business/client/hooks/useBusinessAgentImportMenuItem';
 import { message } from '@/components/AntdStaticMethods';
 import { DESKTOP_HEADER_ICON_SMALL_SIZE } from '@/const/layoutTokens';
+import AgentBreadcrumb from '@/features/AgentBreadcrumb';
 import NavHeader from '@/features/NavHeader';
 import ToggleRightPanelButton from '@/features/RightPanel/ToggleRightPanelButton';
 import { useWorkspaceAwareNavigate } from '@/features/Workspace/useWorkspaceAwareNavigate';
@@ -243,8 +244,12 @@ const Header = memo(() => {
 
   return (
     <NavHeader
+      styles={{ left: { paddingInlineStart: 24 } }}
       left={
         <Flexbox horizontal align={'center'} gap={8}>
+          {activeAgentId && (
+            <AgentBreadcrumb agentId={activeAgentId} title={t('tab.profile', { ns: 'chat' })} />
+          )}
           <AutoSaveHint />
           <AgentStatusTag />
           <AgentVersionReviewTag />
